# Implement `filter` kernel for byte view arrays.